+++ ratings/collection/permissions.js
@@ -1,0 +1,13 @@
+import { ApiBackendRatings } from './';
+
+ApiBackendRatings.allow({
+  insert: function () {
+    // User must be logged in to vote
+    if (Meteor.userId()) {
+      return true;
+    }
+  },
+  update: function () {
+    return true;
+  }
+});
